requirements.txt
langchain==0.2.14
langchain-core==0.2.32
langchain-openai>=0.1.20
pydantic==2.7.4
python-dotenv
ipykernel
langchain_groq
langchain_google_genai
langchain_community
pypdf
faiss-cpu
structlog
PyMuPDF
flask
streamlit
-e.